+++ src/main/java/uz/dckroff/statisfy/dto/notification/DeviceRegistrationRequest.java
@@ -1,0 +1,20 @@
+package uz.dckroff.statisfy.dto.notification;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import javax.validation.constraints.NotBlank;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class DeviceRegistrationRequest {
+
+    @NotBlank(message = "Device token is required")
+    private String token;
+    
+    private String deviceType;
+}